[SYCL][LIBCLC] float16 impl into files responsible for the builtins
In libclc the only way for targets to override the generic
implementation of a builtin is to provide a file that matches the one
from the generic, containing the desired code in target's directory.
With all half floating point builtins implemented in a single file it
was impossible to provide target overrides for selected builtins. This
patch removes `float16.cl` and splits its content across corresponding
builtins files.